EROPSPT-274: Force a lower maximum page size (200)
Signatures are around 20KB on average, so a page size of 500 corresponds
to around 10MB of data, which goes over the limit that can be passed
through an API Gateway.

This change forces at most 200 applications to be returned, even if more
than 200 applications are requested. Requests with a page size over 200
are still accepted, while EMS's are given time to adapt to the change.