EROPSPT-298: Remove index on date created (#89)
This index was originally created to resolve an issues where the sort
buffer could be exhausted with signature data. A different fix was later
added to implement a two-step retrieval process.

We have seen this index used inefficiently for finding ERO downloads on
live - that problem was resolved by adding a new index. We've also seen
this index used when the monitoring queries are running, and the queries
run more slowly when using this index than when forced to use the index
on gss code, status and date created.